+++ integrations/octopus/CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+# Port_Ocean 0.1.75 (2025-04-15)
+
+### Improvements
+
+- Bumped ocean version to ^0.22.4
+
+
+# Port_Ocean 0.1.74 (2025-04-15)
+
+### Improvements
+
+- Bumped ocean version to ^0.22.3
+
+
 # Port_Ocean 0.1.73 (2025-04-07)
 
 ### Improvements